refactor: skeleton 분리
